+++ functorch/functorch/csrc/DynamicLayer.cpp
@@ -6,7 +6,7 @@
 
 #include <functorch/csrc/DynamicLayer.h>
 #include <functorch/csrc/TensorWrapper.h>
-#include <functorch/csrc/BatchedTensorImpl.h>
+#include <ATen/functorch/BatchedTensorImpl.h>
 #include <functorch/csrc/BatchRulesHelper.h>
 
 #include <torch/library.h>
