+++ functorch/functorch/csrc/Interpreter.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <functorch/csrc/Macros.h>
-#include <functorch/csrc/Constants.h>
+#include <ATen/functorch/Constants.h>
 #include <ATen/core/dispatch/Dispatcher.h>
 #include <c10/core/impl/LocalDispatchKeySet.h>
 #include <c10/util/Optional.h>
