+++ functorch/functorch/csrc/PyTorchOperatorHacks.cpp
@@ -1,10 +1,10 @@
 #include <functorch/csrc/DynamicLayer.h>
-#include <functorch/csrc/Constants.h>
+#include <ATen/functorch/Constants.h>
 #include <torch/library.h>
 #include <ATen/ATen.h>
 #include <ATen/WrapDimUtils.h>
 #include <functorch/csrc/TensorWrapper.h>
-#include <functorch/csrc/BatchedTensorImpl.h>
+#include <ATen/functorch/BatchedTensorImpl.h>
 #include <ATen/ATen.h>
 #include <ATen/Dispatch.h>
 #include <c10/util/irange.h>
