+++ functorch/functorch/csrc/VmapModeRegistrations.cpp
@@ -7,9 +7,9 @@
 #include <torch/library.h>
 #include <ATen/ATen.h>
 #include <functorch/csrc/LegacyVmapTransforms.h>
-#include <functorch/csrc/BatchedTensorImpl.h>
+#include <ATen/functorch/BatchedTensorImpl.h>
 #include <functorch/csrc/PlumbingHelper.h>
-#include <functorch/csrc/Constants.h>
+#include <ATen/functorch/Constants.h>
 #include <functorch/csrc/DynamicLayer.h>
 #include <ATen/core/dispatch/Dispatcher.h>
