+++ functorch/functorch/csrc/dim/dim.cpp
@@ -13,7 +13,7 @@
 #include <vector>
 //#include <torch/csrc/autograd/python_variable.h>
 #include <torch/csrc/Export.h>
-#include <functorch/csrc/BatchedTensorImpl.h>
+#include <ATen/functorch/BatchedTensorImpl.h>
 #include <functorch/csrc/DynamicLayer.h>
 #include <ATen/ATen.h>
 #include <memory>
